Address review: move seq-or to V3AssertProp, foreach, UASSERT_OBJ, nonconst delay test
